Add Gems
I’ve identified several gems that will be useful for our tutorial
application.

I learned of these gems from a variety of different sources:

    Ruby Toolbox
    RailsCasts
    RubyFlow
    various blog posts
    example code and starter apps on GitHub
    recommendations from colleagues

We’re adding these gems at the beginning of our development process
since we already know which gems we’ll need. On a real project, you’ll
often discover useful gems and add them to the Gemfile during the
ongoing process of development.

Here are gems we’ll add to the Gemfile:

    activerecord-tableless – helps to use Rails without a database
    foundation-rails – front-end framework
    gibbon – access to the MailChimp API
    google_drive – use Google Drive spreadsheets for data storage
    high_voltage – for static pages like “about”
    simple_form – forms made easy

We’ll also add utilities that make development easier:

    better_errors – helps when things go wrong
    quiet_assets – suppresses distracting messages in the log
    rails_layout – generates files for an application layout

Open your Gemfile and replace the contents with the following:

...