Configure Email
Email messages are visible in the console and the log file when you test
the application. If you don’t want to actually send email, you can skip
this step. But it’s more fun when your application can actually send
email.

You can learn more in the article Send Email with Rails.
Connect to an Email Server

Web servers don’t send email. Our Rails application has to connect to an
email server (also known as a mail transfer agent or “mail relay”). In
the early days of the Internet, an experienced system administrator
could set up an SMTP server to distribute email. Now, because of efforts
to reduce spam, it is necessary to use an established email service to
ensure deliverability. For convenience during development, you can use
your own Gmail account. In production, for high volume transactional
email and improved deliverability, it is best to use a service such as
Mandrill or Mailgun. See the article Send Email with Rails.

We need to configure Rails so the application can connect with an email
server. For our tutorial application, we’ll connect to Gmail to send
email.