Rails Layout Gem with Zurb Foundation
In the previous chapter, we used the rails_layout gem to configure the
default application layout with HTML5 elements, navigation links, and
flash messages. Now we’ll use the rails_layout gem to set up Zurb
Foundation and generate new files for the application layout as well as
the navigation and messages partials. The new files will replace the
layout files we created in the previous chapter.

We’ll use the generator provided by the rails_layout gem uses to set up
Foundation and add the necessary files. Run:

$ rails generate layout:install foundation5 --force

With the --force argument, the rails_layout gem will replace existing
files.