Mailing List
We’ll use the Rails model-view-controller architecture. We’ll need:

    Visitors model
    view for visitors#new
    Visitors controller with new and create methods
    routing for visitors#new and visitors#create

We’ll add a Visitor model that has a data attribute for an email
address. We already have a Visitors controller that renders the home
page using the file in the app/views/visitors/ folder. We’ll replace the
contents of the view file with a nice photo, a marketing message, and a
form.

Our Visitors controller new and create methods will be very similar to
what we created for the Contacts controller. Instead of saving data to
Google Drive, or connecting to Gmail to send a message, we’ll call a
method to save the visitor’s email address to a MailChimp mailing list